STYLE : Mypage QuizList, BookmarkList 컴포넌트 분리
